# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG emissions ingestion and review API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemMetadata } from './activityItemMetadata';
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  description: string;
  /** @nullable */
  actor?: string | null;
  timestamp: Date;
  metadata?: ActivityItemMetadata;
}