# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG emissions ingestion and review API
 * OpenAPI spec version: 0.1.0
 */

export interface AuditEntry {
  id: number;
  /** @nullable */
  recordId?: number | null;
  /** @nullable */
  ingestionId?: number | null;
  action: string;
  /** @nullable */
  actor?: string | null;
  /** @nullable */
  previousValue?: string | null;
  /** @nullable */
  newValue?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}