# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG emissions ingestion and review API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalRecords: number;
  pendingRecords: number;
  approvedRecords: number;
  flaggedRecords: number;
  rejectedRecords: number;
  totalCo2eKg: number;
  approvedCo2eKg: number;
  totalIngestions: number;
  /** Ingestions in the last 7 days */
  recentIngestions: number;
}