# lib/api-zod/src/generated/types/fileUploadInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG emissions ingestion and review API
 * OpenAPI spec version: 0.1.0
 */

export interface FileUploadInput {
  /** Name of the client company */
  clientName?: string;
}