# lib/api-zod/src/generated/types/recordDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Breathe ESG emissions ingestion and review API
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEntry } from './auditEntry';
import type { EmissionRecord } from './emissionRecord';

export type RecordDetail = EmissionRecord & {
  auditTrail: AuditEntry[];
};